chore(user): refactor code
