+++ gradle/libs.versions.toml
@@ -36,7 +36,6 @@ blur = "1.0.7"
 media3 = "1.6.1"
 lrucache = "2.0.2"
 lottie = "6.6.6"
-accompanist = "0.34.0"
 
 [libraries]
 androidx-splash = { module = "androidx.core:core-splashscreen", version.ref = "androidx-splash" }
@@ -85,9 +84,6 @@ mockk-jvm = { module = "io.mockk:mockk-jvm", version.ref = "mockk-jvm" }
 permissions = { module = "com.google.accompanist:accompanist-permissions", version.ref = "permissions" }
 blur = { module = "com.github.Commit451:NativeStackBlur", version.ref = "blur" }
 lottie = { module = "com.airbnb.android:lottie-compose", version.ref = "lottie" }
-accompanist-systemuicontroller = { module = "com.google.accompanist:accompanist-systemuicontroller", version.ref = "accompanist" }
-
-
 
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
